+++ EssentialFeed/EssentialFeediOSTests/FeedViewControllerTests.swift
@@ -438,16 +438,3 @@ private extension UIRefreshControl {
     }
   }
 }
-
-private extension UIImage {
-  static func make(withColor color: UIColor) -> UIImage {
-    let rect = CGRect(x: 0, y: 0, width: 1, height: 1)
-    UIGraphicsBeginImageContext(rect.size)
-    let context = UIGraphicsGetCurrentContext()!
-    context.setFillColor(color.cgColor)
-    context.fill(rect)
-    let img = UIGraphicsGetImageFromCurrentImageContext()
-    UIGraphicsEndImageContext()
-    return img!
-  }
-}
